Multiplayer Logik vollständig (Notifications noch nicht und alle spielen gleichzeitig)

+++ src/main/java/com/blackjack/database/DatabaseLogic.java
@@ -44,7 +44,7 @@ public class DatabaseLogic {
 
 
 
-    public void addUser(String user_name, String user_password, boolean isAdmin, boolean isBanned, float balance) throws SQLException {
+    public void addUser(String user_name, String user_password, boolean isAdmin, boolean isBanned, int balance) throws SQLException {
         if (connection == null) {
             return;
         }
